news clear fix for 2 days and 2 weeks

+++ controllerJobs/news.js
@@ -5,12 +5,19 @@ const { logger } = require('../utils/logger');
 const sequelize = getConnection();
 const removeOutdatedNews = async () => {
   try {
-    const time = getFormattedDaysBefore(2);
-    await sequelize.query(`DELETE from News where createdAt < '${time}'`);
+    let time = getFormattedDaysBefore(2);
+    await sequelize.query(`DELETE from News where type != 'EVT_WEEKLY_ALL_IN_ONE' and createdAt < '${time}'`);
+
+    // clear any message longer than 2 weeks
+    if (new Date().getDay() === 3) {
+      time = getFormattedDaysBefore(16);
+      await sequelize.query(`DELETE from News where createdAt < '${time}'`);
+    }
   } catch (e) {
     logger.error('removeOutdatedNews', e);
   }
 };
+
 
 module.exports = {
   removeOutdatedNews
